backend/requirements.txt
fastapi~=0.103.2
uvicorn~=0.23.2
sqlalchemy~=2.0.20
pyodbc~=4.0.39
passlib[bcrypt]~=1.7.4
python-jose~=3.3.0
pydantic>=2.0.0,<3.0.0
python-multipart~=0.0.6
gunicorn~=21.2.0
python-dotenv~=1.0.0
fastapi-mail~=1.4.1
sendgrid~=6.10.0
slowapi~=0.1.8
jinja2~=3.1.2
email-validator~=2.0.0
bcrypt~=4.0.1